+++ packages/plugin-ai-video-generation-web/src/fal-ai/index.ts
@@ -4,6 +4,8 @@ import { KlingVideoV21MasterImageToVideo } from './KlingVideoV21MasterImageToVid
 import { PixverseV35TextToVideo } from './PixverseV35TextToVideo';
 import { KlingVideoV21MasterTextToVideo } from './KlingVideoV21MasterTextToVideo';
 import { Veo3TextToVideo } from './Veo3TextToVideo';
+import { ByteDanceSeedanceV1ProImageToVideo } from './ByteDanceSeedanceV1ProImageToVideo';
+import { ByteDanceSeedanceV1ProTextToVideo } from './ByteDanceSeedanceV1ProTextToVideo';
 
 const FalAi = {
   MinimaxVideo01Live,
@@ -11,7 +13,9 @@ const FalAi = {
   KlingVideoV21MasterImageToVideo,
   PixverseV35TextToVideo,
   KlingVideoV21MasterTextToVideo,
-  Veo3TextToVideo
+  Veo3TextToVideo,
+  ByteDanceSeedanceV1ProImageToVideo,
+  ByteDanceSeedanceV1ProTextToVideo
 };
 
 export default FalAi;